(PCP-741) Update tk-jetty and jetty-websockets
Error handling during the intial upgrade request changed for this
version of jetty.
1. Errors that happen before the connection is upgraded to a websocket
are now wrapped in UpgradeException. This commit unwraps and rethrows
them so we can catch some specific connection errors.
2. A connection error results in two errors being passed to the
Listener's onWebsocketError handler. gnaidzo throws the first on of
these synchonously from it's on connect function but we have to handle
the second one to avoid uneccesarry logging.