(maint) Update messaging_test regex
This commit updates the TLS Handshake regex to account for a possible
race condition. It seems possible for the "Sleeping for up to 200 ms"
message can occur without being logged at warn level (from a regular
connection failure), so the first message in the logs has the 400 ms
timeout instead. This commit should help avoid that issue.