(PE-34653) update clj-parent to 5.2.9 remove stylefruits pin
This updates the clj-parent version to 5.2.9, which contains
a version declaration for stylefruits/gniazdo, allowing the version
in this repository to not pin a specific version.

The release series was bumped up to 1.4.x due to the changes in
clj-parent.

The CHANGELOG was also updated in anticipation of a release.