(maint) add certs to Client record
Recently the additional `certs` entry was needed in the `Client` record,
but wasn't added to the record definition.  If schema validation is turned
on, this causes an exception to be thrown.

This updates the record to contain the certs entry, adds the schema
validation to both test namespaces, and removes an unused parameter.

A pointless test was removed as it violated the schema result of
the function, and creating a valid payload would be more effort
than the test is worth.